Refactor to join `QueryableSystem` implementations
Create a new `impl_queryable_system` macro that is able to generate the
code for implementing `QueryableSystem` for a specified Service system
API type. Change the `wasmer` and `wasmtime` modules to use the new
macro.
