+++ linera-execution/src/wasm/system_api.rs
@@ -292,3 +292,167 @@ macro_rules! impl_writable_system {
         }
     };
 }
+
+/// Generates an implementation of `QueryableSystem` for the provided `service_system_api` type.
+///
+/// Generates the common code for service system API types for all WASM runtimes.
+macro_rules! impl_queryable_system {
+    ($service_system_api:ident<$storage:lifetime>) => {
+        impl_queryable_system!(@generate $service_system_api<$storage>, $storage, <$storage>);
+    };
+
+    ($service_system_api:ident) => {
+        impl_queryable_system!(@generate $service_system_api, 'static);
+    };
+
+    (@generate $service_system_api:ty, $storage:lifetime $(, <$param:lifetime> )?) => {
+        impl$(<$param>)? QueryableSystem for $service_system_api {
+            type Load = HostFuture<$storage, Result<Vec<u8>, ExecutionError>>;
+            type Lock = HostFuture<$storage, Result<(), ExecutionError>>;
+            type Unlock = HostFuture<$storage, Result<(), ExecutionError>>;
+            type ReadKeyBytes = HostFuture<$storage, Result<Option<Vec<u8>>, ExecutionError>>;
+            type FindKeys = HostFuture<$storage, Result<Vec<Vec<u8>>, ExecutionError>>;
+            type FindKeyValues =
+                HostFuture<$storage, Result<Vec<(Vec<u8>, Vec<u8>)>, ExecutionError>>;
+            type TryQueryApplication = HostFuture<$storage, Result<Vec<u8>, ExecutionError>>;
+
+            fn chain_id(&mut self) -> queryable_system::ChainId {
+                self.storage().chain_id().into()
+            }
+
+            fn application_id(&mut self) -> queryable_system::ApplicationId {
+                self.storage().application_id().into()
+            }
+
+            fn application_parameters(&mut self) -> Vec<u8> {
+                self.storage().application_parameters()
+            }
+
+            fn read_system_balance(&mut self) -> queryable_system::SystemBalance {
+                self.storage().read_system_balance().into()
+            }
+
+            fn read_system_timestamp(&mut self) -> queryable_system::Timestamp {
+                self.storage().read_system_timestamp().micros()
+            }
+
+            fn load_new(&mut self) -> Self::Load {
+                HostFuture::new(self.storage().try_read_my_state())
+            }
+
+            fn load_poll(&mut self, future: &Self::Load) -> queryable_system::PollLoad {
+                use queryable_system::PollLoad;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollLoad::Pending,
+                    Poll::Ready(Ok(bytes)) => PollLoad::Ready(Ok(bytes)),
+                    Poll::Ready(Err(error)) => PollLoad::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn lock_new(&mut self) -> Self::Lock {
+                HostFuture::new(self.storage().lock_view_user_state())
+            }
+
+            fn lock_poll(&mut self, future: &Self::Lock) -> queryable_system::PollLock {
+                use queryable_system::PollLock;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollLock::Pending,
+                    Poll::Ready(Ok(())) => PollLock::Ready(Ok(())),
+                    Poll::Ready(Err(error)) => PollLock::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn unlock_new(&mut self) -> Self::Unlock {
+                HostFuture::new(self.storage().unlock_view_user_state())
+            }
+
+            fn unlock_poll(&mut self, future: &Self::Lock) -> queryable_system::PollUnlock {
+                use queryable_system::PollUnlock;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollUnlock::Pending,
+                    Poll::Ready(Ok(())) => PollUnlock::Ready(Ok(())),
+                    Poll::Ready(Err(error)) => PollUnlock::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn read_key_bytes_new(&mut self, key: &[u8]) -> Self::ReadKeyBytes {
+                HostFuture::new(self.storage().read_key_bytes(key.to_owned()))
+            }
+
+            fn read_key_bytes_poll(
+                &mut self,
+                future: &Self::ReadKeyBytes,
+            ) -> queryable_system::PollReadKeyBytes {
+                use queryable_system::PollReadKeyBytes;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollReadKeyBytes::Pending,
+                    Poll::Ready(Ok(opt_list)) => PollReadKeyBytes::Ready(Ok(opt_list)),
+                    Poll::Ready(Err(error)) => PollReadKeyBytes::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn find_keys_new(&mut self, key_prefix: &[u8]) -> Self::FindKeys {
+                HostFuture::new(self.storage().find_keys_by_prefix(key_prefix.to_owned()))
+            }
+
+            fn find_keys_poll(&mut self, future: &Self::FindKeys) -> queryable_system::PollFindKeys {
+                use queryable_system::PollFindKeys;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollFindKeys::Pending,
+                    Poll::Ready(Ok(keys)) => PollFindKeys::Ready(Ok(keys)),
+                    Poll::Ready(Err(error)) => PollFindKeys::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn find_key_values_new(&mut self, key_prefix: &[u8]) -> Self::FindKeyValues {
+                HostFuture::new(
+                    self.storage()
+                        .find_key_values_by_prefix(key_prefix.to_owned()),
+                )
+            }
+
+            fn find_key_values_poll(
+                &mut self,
+                future: &Self::FindKeyValues,
+            ) -> queryable_system::PollFindKeyValues {
+                use queryable_system::PollFindKeyValues;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollFindKeyValues::Pending,
+                    Poll::Ready(Ok(key_values)) => PollFindKeyValues::Ready(Ok(key_values)),
+                    Poll::Ready(Err(error)) => PollFindKeyValues::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn try_query_application_new(
+                &mut self,
+                application: queryable_system::ApplicationId,
+                argument: &[u8],
+            ) -> Self::TryQueryApplication {
+                let storage = self.storage();
+                let argument = Vec::from(argument);
+
+                HostFuture::new(async move {
+                    storage
+                        .try_query_application(application.into(), &argument)
+                        .await
+                })
+            }
+
+            fn try_query_application_poll(
+                &mut self,
+                future: &Self::TryQueryApplication,
+            ) -> queryable_system::PollLoad {
+                use queryable_system::PollLoad;
+                match future.poll(self.context()) {
+                    Poll::Pending => PollLoad::Pending,
+                    Poll::Ready(Ok(result)) => PollLoad::Ready(Ok(result)),
+                    Poll::Ready(Err(error)) => PollLoad::Ready(Err(error.to_string())),
+                }
+            }
+
+            fn log(&mut self, message: &str, level: queryable_system::LogLevel) {
+                log::log!(level.into(), "{message}");
+            }
+        }
+    };
+}
